+++ app/bootstrap.php
@@ -4,6 +4,11 @@ declare(strict_types=1);
 
 require __DIR__ . '/support/helpers.php';
 
+$composerAutoload = dirname(__DIR__) . '/vendor/autoload.php';
+if (is_file($composerAutoload)) {
+    require $composerAutoload;
+}
+
 spl_autoload_register(function (string $class): void {
     $prefix = 'App\\';
     $baseDir = __DIR__ . '/';